coffebot/requirements.txt
django
South
django-choices
twitter
django-singletons
django-bootstrapped
simplejson
django_chartit
